+++ src/main/kotlin/icu/windea/pls/script/codeInsight/hints/ParadoxDefinitionReferenceLocalizedNameHintsProvider.kt
@@ -89,6 +89,13 @@ class ParadoxDefinitionReferenceLocalizedNameHintsProvider : ParadoxScriptHintsP
 					?: return true
 				CwtConfigHandler.resolveScriptExpression(element, null, config, config.info.configGroup, false)
 			}
+			//这也是需要判断的
+			is ParadoxScriptInt -> {
+				val config = ParadoxCwtConfigHandler.resolveValueConfigs(element).firstOrNull()
+					?.takeIf { it.expression.type in valueExpressionTypes }
+					?: return true
+				CwtConfigHandler.resolveScriptExpression(element, null, config, config.info.configGroup, false)
+			}
 			else -> return true
 		}
 		if(resolved is ParadoxScriptDefinitionElement) {